pyproject.toml
[project]
name = "coderr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.1",
    "django-cors-headers>=4.9.0",
    "django-extensions>=4.1",
    "djangorestframework>=3.16.1",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = ["ruff>=0.14.14"]

[tool.ruff]
line-length = 79
target-version = "py38"
indent-width = 4

exclude = [".git", "__pycache__", "build", "dist", "venv"]

[tool.ruff.lint]
select = ["E", "F"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
